orderAPI

+++ Rocoland/Models/OrderItem.cs
@@ -4,7 +4,7 @@
 
         public int Id { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public Product Product { get; set; }
         public int ProductId { get; set; }
         public virtual Order Order { get; set; }
